requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0
dj-database-url>=2.1,<3.0
